+++ src/main/scala/is/hail/io/plink/ExportPlink.scala
@@ -1,12 +1,16 @@
 package is.hail.io.plink
 
 import java.io.OutputStreamWriter
+
+import is.hail.HailContext
 import is.hail.annotations.Region
+import is.hail.expr.MatrixValue
 import is.hail.expr.types._
 import is.hail.variant._
 import is.hail.utils._
+import org.apache.spark.TaskContext
 
-object ExportPlinkUtils {
+object ExportPlink {
   val bedHeader = Array[Byte](108, 27, 1)
   val gtMap = Array(3, 2, 0)
   val spaceRegex = """\s+""".r
@@ -51,6 +55,67 @@ object ExportPlinkUtils {
       k += 1
     }
     bp.flush()
+  }
+
+  def apply(mv: MatrixValue, path: String): Unit = {
+    val hc = HailContext.get
+    val sc = hc.sc
+    val hConf = hc.hadoopConf
+
+    val tmpBedDir = hConf.getTemporaryFile(hc.tmpDir)
+    val tmpBimDir = hConf.getTemporaryFile(hc.tmpDir)
+
+    hConf.mkDir(tmpBedDir)
+    hConf.mkDir(tmpBimDir)
+
+    val sHConfBc = sc.broadcast(new SerializableHadoopConfiguration(hConf))
+
+    val nPartitions = mv.rvd.getNumPartitions
+    val d = digitsNeeded(nPartitions)
+
+    val nSamples = mv.colValues.value.length
+    val fullRowType = mv.typ.rvRowType
+
+    val nRecordsWritten = mv.rvd.mapPartitionsWithIndex { case (i, it) =>
+      val hConf = sHConfBc.value.value
+      val f = partFile(d, i, TaskContext.get)
+      val bedPartPath = tmpBedDir + "/" + f
+      val bimPartPath = tmpBimDir + "/" + f
+      var rowCount = 0L
+
+      hConf.writeTextFile(bimPartPath) { bimOS =>
+        hConf.writeFile(bedPartPath) { bedOS =>
+          val v = new RegionValueVariant(fullRowType)
+          val a = new BimAnnotationView(fullRowType)
+          val hcv = HardCallView(fullRowType)
+          val bp = new BitPacker(2, (i: Int) => bedOS.write(i))
+
+          it.foreach { rv =>
+            v.setRegion(rv)
+            a.setRegion(rv)
+            ExportPlink.writeBimRow(v, a, bimOS)
+
+            hcv.setRegion(rv)
+            ExportPlink.writeBedRow(hcv, bp, nSamples)
+
+            rowCount += 1
+          }
+        }
+      }
+
+      Iterator.single(rowCount)
+    }.collect().sum
+
+    hConf.writeFile(tmpBedDir + "/_SUCCESS")(out => ())
+    hConf.writeFile(tmpBedDir + "/header")(out => out.write(ExportPlink.bedHeader))
+    hConf.copyMerge(tmpBedDir, path + ".bed", nPartitions, header = true)
+
+    hConf.writeTextFile(tmpBimDir + "/_SUCCESS")(out => ())
+    hConf.copyMerge(tmpBimDir, path + ".bim", nPartitions, header = false)
+
+    mv.colsTableValue.export(path + ".fam", header = false)
+
+    info(s"wrote $nRecordsWritten variants and $nSamples samples to '$path'")
   }
 }
